Fix context leak in prepareStatement
prepareStatement deduplicates in-flight requests for preparing the same
statement, however it failed the preparation in case the context passed
from the first goroutine requesting was canceled. prepareStatment
returned context.Canceled error to all goroutines, even if their context
was not canceled, so the context from first goroutine leaked to all
others..

We need to move the prepare statement call to a separate goroutine so
that it can continue even if the first goroutine that spawned the task
does not want to wait anymore, since there might be other goroutines
waiting for the prepareStatement to complete.