Add TokenAwareLatencyHostPolicy
This policy should better handle network issues where latency of some
subset of nodes is worse (e.g. when packets are dropped and the data
needs to be retransmitted). Tracking latency allows us to detect this
case and switch to different replicas instead.

This commit introduces dependencies on github.com/klauspost/cpuid/v2
and golang.org/x/sys, so that we can use per-CPU data structures for
recording latency. This is necessary to avoid lock contention as
many queries need to write latency data to the same hosts.

Co-authored-by: Matouš Dzivjak <matous.dzivjak@kiwi.com>
Co-authored-by: Ivan Boyarkin <ivan.boyarkin@kiwi.com>