Fix call leak in Conn.exec
We see

   attempting to use stream already in use

errors in logs. This seems to be introduced by
1565c9efe9edf70474ef4b2b07c79da96ec8e5f1 that added cancellable
write to the connection.

Indeed the in case the frame is too big or the context is canceled
before we start writing to the connection, we didn't properly release
the call. We only released the stream, which allows it to be returned
by GetStream(), but we left the call in place, which would always
cause the attempting to use stream already in use error for that
specific stream id. Normally the call is released once we get reply
from the connection, but in those cases we don't write so the
reply from the connection never happens.