Don't hold connection mutex when closing callReqs
Previously when closing the connection, closeWithError would send
connection close errors to all currently outstanding callReqs while
holding the connection mutex. During this time, other goroutines trying
to start a request on this connection are blocked. These goroutines
could have created new callReqs for this connection and started trying
to write. The write would be probably canceled by closing the
underlying connection.

This commit changes the behaviour so that connection is marked as closed
and no new callReqs are created for the connection nor any more frames
are read. Taking all the callReqs to the goroutine that is closing
the connection allows us to iterate the callReqs without holding the
connection lock. If the iteration takes long (or if there is a
deadlock), only one goroutine is affected instead of any goroutine that
would try to create a request on this connection.