Make sure policies are not reused between sessions
Sharing a policy between sessions is not supported because the policy
receives state updates from the session.

Let's update the documentation and add a panic in TokenAwareHostPolicy
constructor. It is better to panic early than to have races and
undefined behavior later.

See also discussion in https://github.com/scylladb/gocql/issues/94